requirements.txt
fastapi==0.115.6
openai==1.57.0
pydantic==2.10.3
python-dotenv==1.0.1
replicate==1.0.4
Requests==2.32.3
